Add timeoutconn
timeoutconn is a wrapper around a net.Conn which enforces a timeout.
The timeout will be reset on every read or write operation.

This is not the same as setting a deadline. If a deadline is set then a
conn will only get closed on the next read or write operation, otherwise
it will linger.